moved functions to utils/view.py
