+++ pipeline/views.py
@@ -26,7 +26,7 @@ from django.contrib.postgres.aggregates.general import ArrayAgg
 from django.contrib.auth.decorators import login_required
 
 
-from .models import Image, Measurement, Run, Source, SkyRegion
+from .models import Image, Measurement, Run, Source
 from .serializers import (
     ImageSerializer, MeasurementSerializer, RunSerializer,
     SourceSerializer
@@ -34,83 +34,10 @@ from .serializers import (
 from .utils.utils import (
     deg2dms, deg2hms, gal2equ, ned_search, simbad_search
 )
-from .utils.view import FLOAT_FIELDS
+from .utils.view import generate_colsfields, get_skyregions_collection
 
 
 logger = logging.getLogger(__name__)
-
-
-def generate_colsfields(fields, url_prefix):
-    colsfields = []
-
-    for col in fields:
-        if col == 'name':
-            colsfields.append({
-                'data': col, 'render': {
-                    'url': {
-                        'prefix': url_prefix,
-                        'col': 'name'
-                    }
-                }
-            })
-        elif col in FLOAT_FIELDS:
-            colsfields.append({
-                'data': col,
-                'render': {
-                    'float': {
-                        'col': col,
-                        'precision': FLOAT_FIELDS[col]['precision'],
-                        'scale': FLOAT_FIELDS[col]['scale'],
-                    }
-                }
-            })
-        else:
-            colsfields.append({'data': col})
-
-    return colsfields
-
-
-def get_skyregions_collection():
-    """
-    Produce Sky region geometry shapes for d3-celestial.
-    """
-    skyregions = SkyRegion.objects.all()
-
-    features = []
-
-    for skr in skyregions:
-        ra = skr.centre_ra - 180.
-        dec = skr.centre_dec
-        width_ra = skr.width_ra / 2.
-        width_dec = skr.width_dec / 2.
-        id = skr.id
-        features.append(
-            {
-                "type": "Feature",
-                "id": f"SkyRegion{id}",
-                "properties": {
-                    "n": f"{id:02d}",
-                    "loc": [ra, dec]
-                },
-                "geometry": {
-                    "type": "MultiLineString",
-                    "coordinates": [[
-                        [ra+width_ra, dec+width_dec],
-                        [ra+width_ra, dec-width_dec],
-                        [ra-width_ra, dec-width_dec],
-                        [ra-width_ra, dec+width_dec],
-                        [ra+width_ra, dec+width_dec]
-                    ]]
-                }
-            }
-        )
-
-    skyregions_collection = {
-        "type": "FeatureCollection",
-        "features" : features
-    }
-
-    return skyregions_collection
 
 
 def Login(request):
